@@ -3,8 +3,8 @@ package com.airbnb.deeplinkdispatch.sample.library;
 import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.Toast;
 
 import com.airbnb.deeplinkdispatch.DeepLink;
